+++ packages/ui/src/copy-to-clipboard/index.ts
@@ -1,0 +1,1 @@
+export { CopyToClipboard } from './copy-to-clipboard';